dom/tests/mochitest/general/test_donottrack.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=629535
-->
<head>
  <title>Test for Bug 629535</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=629535">Mozilla Bug 629535</a>

<script type="application/javascript">

const dntPref = 'privacy.donottrackheader.enabled';
const dntValuePref = 'privacy.donottrackheader.value';
const oldDNT = SpecialPowers.getBoolPref(dntPref);
const oldDNTValue = SpecialPowers.getIntPref(dntValuePref);

is(SpecialPowers.getBoolPref(dntPref), false,
   'DNT should be disabled by default');
is(SpecialPowers.getIntPref(dntValuePref), "1",
   'DNT value should be "1" by default');
is(navigator.doNotTrack, 'unspecified',
   'navigator.doNotTrack should initially be "unspecified".');

SpecialPowers.clearUserPref(dntPref);
is(navigator.doNotTrack, "unspecified", 'after clearing pref');

SpecialPowers.setBoolPref(dntPref, true);
is(navigator.doNotTrack, "1", 'after setting pref to true');

SpecialPowers.clearUserPref(dntValuePref);
is(navigator.doNotTrack, "1", 'after clearing value pref');

SpecialPowers.setIntPref(dntValuePref, "0");
is(navigator.doNotTrack, "0", 'after setting value pref');

SpecialPowers.setIntPref(dntValuePref, "42");
is(navigator.doNotTrack, "1", 'after setting value pref');

SpecialPowers.clearUserPref(dntValuePref);
is(navigator.doNotTrack, "1", 'after clearing value pref');

SpecialPowers.setBoolPref(dntPref, false);
is(navigator.doNotTrack, "unspecified", 'after setting pref to false');

SpecialPowers.setBoolPref(dntPref, oldDNT);
SpecialPowers.setIntPref(dntValuePref, oldDNTValue);

</script>

</body>
</html>
